pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rafael Fonseca <rfonseca85@yahoo.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic = "^2.7.0"
openai = "^1.23.2"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pymongo = "^4.6.3"
python-dotenv = "^1.0.1"
bcrypt = "4.0.1"
python-docx = "^1.1.0"
websockets = "^12.0"
llm = "^0.13.1"
docxtpl = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
